Move all Storage implementations in storage.py
